refactor: remove componente TableWrapperWithStyle
