--- src/features/acf/frontend/hypertension/modules/AcfPage/modules/List/modules/common/RenderPatientNameCpfCns/modules/TableWrapperWithStyle.tsx
@@ -1,19 +1,0 @@
-import { Box } from "@mui/material";
-
-export const TableWrapperWithStyle: React.FC<React.PropsWithChildren> = ({
-    children,
-}) => {
-    return (
-        <Box
-            sx={{
-                "& .patient-name-cpf-cns": {
-                    whiteSpace: "break-spaces !important",
-                    paddingTop: "8px !important",
-                    paddingBottom: "8px !important",
-                },
-            }}
-        >
-            {children}
-        </Box>
-    );
-};
